PracticePlatform/Scheduling/backend/requirements.txt
# Scheduling2.0 - Backend Dependencies
# Merged requirements from both MBH-Scheduling and Scheduling projects

# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Email
fastapi-mail==1.4.1

# Date/Time
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# AI Agent Dependencies
langchain==0.1.0
langchain-core>=0.1.7,<0.2
langgraph==0.0.20
openai==1.3.0
anthropic==0.7.0 